# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Summer Learning Adventure API
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface Profile {
  id: number;
  name: string;
  avatarEmoji: string;
  totalStars: number;
  currentStreak: number;
  createdAt: string;
}

export interface UpdateProfileBody {
  name: string;
  avatarEmoji: string;
}

export interface Category {
  id: number;
  slug: string;
  name: string;
  description: string;
  emoji: string;
  colorHex: string;
  totalLessons: number;
  completedLessons: number;
}

export type LessonDifficulty =
  (typeof LessonDifficulty)[keyof typeof LessonDifficulty];

export const LessonDifficulty = {
  beginner: "beginner",
  intermediate: "intermediate",
  advanced: "advanced",
} as const;

export interface Lesson {
  id: number;
  categoryId: number;
  title: string;
  week: number;
  dayOrder: number;
  difficulty: LessonDifficulty;
  totalActivities: number;
  completedActivities: number;
  starsEarned: number;
  isUnlocked: boolean;
}

export type LessonWithCategory = Lesson & {
  categoryName: string;
  categoryEmoji: string;
  categoryColorHex: string;
  categorySlug: string;
};

export type ActivityType = (typeof ActivityType)[keyof typeof ActivityType];

export const ActivityType = {
  sight_word_flash: "sight_word_flash",
  spelling_drag: "spelling_drag",
  rhyme_match: "rhyme_match",
  counting_tap: "counting_tap",
  pattern_fill: "pattern_fill",
  reading_passage: "reading_passage",
  addition_basic: "addition_basic",
  letter_match: "letter_match",
} as const;

export interface Activity {
  id: number;
  lessonId: number;
  type: ActivityType;
  orderIndex: number;
  /** JSON-encoded activity data (words, images, options, correct answers) */
  contentJson: string;
  isCompleted: boolean;
  starsEarned: number;
}

export type LessonDetail = Lesson & {
  activities: Activity[];
};

export interface ActivityProgress {
  id: number;
  activityId: number;
  starsEarned: number;
  completedAt: string;
}

export interface RecordProgressBody {
  activityId: number;
  starsEarned: number;
}

export interface CategoryProgress {
  categoryId: number;
  categoryName: string;
  emoji: string;
  colorHex: string;
  completedLessons: number;
  totalLessons: number;
  starsEarned: number;
}

export interface StatsSummary {
  totalStars: number;
  totalActivitiesCompleted: number;
  totalLessonsCompleted: number;
  currentStreak: number;
  currentWeek: number;
  totalWeeks: number;
  categoriesProgress: CategoryProgress[];
}

export interface Achievement {
  id: number;
  slug: string;
  name: string;
  description: string;
  emoji: string;
  isUnlocked: boolean;
  unlockedAt?: string | null;
}

export interface WeekSummary {
  week: number;
  activitiesCompleted: number;
  starsEarned: number;
  lessonsCompleted: number;
}