# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: Summer Learning Adventure API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /categories/{categoryId}/lessons:
    get:
      operationId: listLessons
      tags: [lessons]
      parameters:
        - name: categoryId
          in: path
          required: true
          schema:
            type: integer
      summary: List all lessons in a category
      responses:
        "200":
          description: List of lessons
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Lesson"
  /lessons/{lessonId}:
    get:
      operationId: getLesson
      tags: [lessons]
      parameters:
        - name: lessonId
          in: path
          required: true
          schema:
            type: integer
      summary: Get a lesson with all its activities
      responses:
        "200":
          description: Lesson with activities
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/LessonDetail"
...
components:
  schemas:
...
    Lesson:
      type: object
      properties:
        id:
          type: integer
        categoryId:
          type: integer
        title:
          type: string
        week:
          type: integer
        dayOrder:
          type: integer
        difficulty:
          type: string
          enum: [beginner, intermediate, advanced]
        totalActivities:
          type: integer
        completedActivities:
          type: integer
        starsEarned:
          type: integer
        isUnlocked:
          type: boolean
      required: [id, categoryId, title, week, dayOrder, difficulty, totalActivities, completedActivities, starsEarned, isUnlocked]
...
    Activity:
      type: object
      properties:
        id:
          type: integer
        lessonId:
          type: integer
        type:
          type: string
          enum: [sight_word_flash, spelling_drag, rhyme_match, counting_tap, pattern_fill, reading_passage, addition_basic, letter_match]
        orderIndex:
          type: integer
        contentJson:
          type: string
          description: JSON-encoded activity data (words, images, options, correct answers)
        isCompleted:
          type: boolean
        starsEarned:
          type: integer
      required: [id, lessonId, type, orderIndex, contentJson, isCompleted, starsEarned]
    LessonDetail:
      allOf:
        - $ref: "#/components/schemas/Lesson"
        - type: object
          properties:
            activities:
              type: array
              items:
                $ref: "#/components/schemas/Activity"
          required: [activities]
...